Extract with and anonymous wrapping fixtures
